refactor: change function createCars
